Add full and small logo in AppNav
